+++ packages/talker/lib/src/addons/addons.dart
@@ -1,0 +1,1 @@
+export 'original_addons.dart';