+++ packages/talker/lib/src/talker.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+
 import 'package:talker/talker.dart';
 
 /// Talker - advanced exception handling and logging
@@ -37,7 +38,7 @@ class Talker implements TalkerInterface {
     Function(String message)? loggerOutput,
   }) {
     _filter = filter;
-    _settings = settings ?? TalkerSettings();
+    this.settings = settings ?? TalkerSettings();
     _logger = logger ??
         TalkerLogger().copyWith(
           settings: loggerSettings,
@@ -48,11 +49,14 @@ class Talker implements TalkerInterface {
     if (observers != null && observers.isNotEmpty) {
       _observersManager = TalkerObserversManager(observers);
     }
-    _errorHandler = TalkerErrorHandler(_settings);
+    _errorHandler = TalkerErrorHandler(this.settings);
   }
 
   /// Fields can be setup in [configure()] method
-  late TalkerSettings _settings;
+
+  /// {@macro talker_settings}
+  @override
+  late TalkerSettings settings;
   late TalkerLoggerInterface _logger;
   late TalkerErrorHandlerInterface _errorHandler;
   late TalkerFilter? _filter;
@@ -60,6 +64,7 @@ class Talker implements TalkerInterface {
   // final _fileManager = FileManager();
   final _history = <TalkerDataInterface>[];
   TalkerObserversManager? _observersManager;
+  final _addons = <String, Object>{};
 
   /// {@macro talker_configure}
   @override
@@ -76,7 +81,7 @@ class Talker implements TalkerInterface {
       _filter = filter;
     }
     if (settings != null) {
-      _settings = settings;
+      this.settings = settings;
     }
 
     if (observers != null && observers.isNotEmpty) {
@@ -106,6 +111,10 @@ class Talker implements TalkerInterface {
   /// {@macro talker_history}
   @override
   List<TalkerDataInterface> get history => _history;
+
+  ///{@macro talker_addons}
+  @override
+  Map<String, Object> get addons => _addons;
 
   /// {@macro talker_handle}
   @override
@@ -144,7 +153,7 @@ class Talker implements TalkerInterface {
       logLevel: LogLevel.error,
     );
     _handleErrorData(data);
-    if (_settings.enabled) {
+    if (settings.enabled) {
       _observersManager?.onError(data);
     }
   }
@@ -164,7 +173,7 @@ class Talker implements TalkerInterface {
       logLevel: LogLevel.error,
     );
     _handleErrorData(data);
-    if (_settings.enabled) {
+    if (settings.enabled) {
       _observersManager?.onException(data);
     }
   }
@@ -270,7 +279,7 @@ class Talker implements TalkerInterface {
   ///{@macro talker_clear_log_history}
   @override
   void cleanHistory() {
-    if (_settings.useHistory) {
+    if (settings.useHistory) {
       _history.clear();
     }
   }
@@ -278,13 +287,31 @@ class Talker implements TalkerInterface {
   ///{@macro talker_disable}
   @override
   void disable() {
-    _settings.enabled = false;
+    settings.enabled = false;
   }
 
   ///{@macro talker_enable}
   @override
   void enable() {
-    _settings.enabled = true;
+    settings.enabled = true;
+  }
+
+  ///{@macro talker_addon_register}
+  @override
+  void registerAddon({
+    required String code,
+    required Object addon,
+  }) {
+    if (_addons.containsKey(code)) {
+      throw Exception('Addon currently exist');
+    }
+    _addons[code] = addon;
+  }
+
+  ///{@macro talker_addon_reset}
+  @override
+  void resetAddon(String code) {
+    _addons.remove(code);
   }
 
   void _handleLog(
@@ -305,7 +332,7 @@ class Talker implements TalkerInterface {
   }
 
   void _handleErrorData(TalkerDataInterface data) {
-    if (!_settings.enabled) {
+    if (!settings.enabled) {
       return;
     }
     final isApproved = _isApprovedByFilter(data);
@@ -314,7 +341,7 @@ class Talker implements TalkerInterface {
     }
     _talkerStreamController.add(data);
     _handleForOutputs(data);
-    if (_settings.useConsoleLogs) {
+    if (settings.useConsoleLogs) {
       _logger.log(
         data.generateTextMessage(),
         level: data.logLevel ?? LogLevel.error,
@@ -327,7 +354,7 @@ class Talker implements TalkerInterface {
     AnsiPen? pen,
     LogLevel? logLevel,
   }) {
-    if (!_settings.enabled) {
+    if (!settings.enabled) {
       return;
     }
     final isApproved = _isApprovedByFilter(data);
@@ -337,7 +364,7 @@ class Talker implements TalkerInterface {
     _observersManager?.onLog(data);
     _talkerStreamController.add(data);
     _handleForOutputs(data);
-    if (_settings.useConsoleLogs) {
+    if (settings.useConsoleLogs) {
       _logger.log(
         data.generateTextMessage(),
         level: logLevel ?? data.logLevel,
@@ -359,8 +386,8 @@ class Talker implements TalkerInterface {
   // }
 
   void _writeToHistory(TalkerDataInterface data) {
-    if (_settings.useHistory && _settings.enabled) {
-      if (_settings.maxHistoryItems <= _history.length) {
+    if (settings.useHistory && settings.enabled) {
+      if (settings.maxHistoryItems <= _history.length) {
         _history.removeAt(0);
       }
       _history.add(data);
@@ -370,5 +397,24 @@ class Talker implements TalkerInterface {
   bool _isApprovedByFilter(TalkerDataInterface data) {
     final approved = _filter?.filter(data) ?? true;
     return approved;
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+
+    return other is Talker &&
+        other.settings == settings &&
+        other._logger == _logger &&
+        other._errorHandler == _errorHandler &&
+        other._filter == _filter;
+  }
+
+  @override
+  int get hashCode {
+    return settings.hashCode ^
+        _logger.hashCode ^
+        _errorHandler.hashCode ^
+        _filter.hashCode;
   }
 }