+++ packages/talker/lib/src/utils/utils.dart
@@ -1,4 +1,3 @@
 export 'error_handler/error_handler.dart';
-export 'file_manager/file_manager.dart';
 export 'filter/filter.dart';
 export 'formaters/formaters.dart';