+++ packages/talker/test/talker_settings_test.dart
@@ -25,6 +25,52 @@ void main() {
         true,
       );
     });
+
+    test('Equality', () async {
+      final settings = TalkerSettings(
+        useConsoleLogs: false,
+      );
+
+      final settings2 = TalkerSettings(
+        useConsoleLogs: false,
+      );
+
+      expect(settings, settings2);
+    });
+
+    test('hashCode', () async {
+      final settings = TalkerSettings(
+        useConsoleLogs: false,
+      );
+
+      expect(settings.hashCode, isNotNull);
+      expect(settings.hashCode, isNot(0));
+    });
+
+    test('copyWith', () async {
+      final settings = TalkerSettings();
+      final newSettings = settings.copyWith(
+        enabled: false,
+        useHistory: false,
+        useConsoleLogs: false,
+        maxHistoryItems: 999,
+      );
+
+      expect(newSettings.enabled, false);
+      expect(newSettings.useConsoleLogs, false);
+      expect(newSettings.useHistory, false);
+      expect(newSettings.maxHistoryItems, 999);
+    });
+
+    test('copyWith empty', () async {
+      final settings = TalkerSettings();
+      final newSettings = settings.copyWith();
+
+      expect(newSettings.enabled, true);
+      expect(newSettings.useConsoleLogs, true);
+      expect(newSettings.useHistory, true);
+      expect(newSettings.maxHistoryItems, 200);
+    });
   });
 }