+++ packages/talker_bloc_logger/lib/src/talker_bloc_logger_settings.dart
@@ -1,0 +1,11 @@
+class TalkerBlocLoggerSettings {
+  const TalkerBlocLoggerSettings({
+    this.enabled = true,
+    this.printEventFullData = true,
+    this.printStateFullData = true,
+  });
+
+  final bool enabled;
+  final bool printEventFullData;
+  final bool printStateFullData;
+}