+++ packages/talker_flutter/lib/src/models/additional_talker_setting.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class AdditionalTalkerSetting {
+  const AdditionalTalkerSetting({
+    required this.title,
+    required this.value,
+    required this.onChanged,
+  });
+
+  final String title;
+  final bool value;
+  final ValueChanged<bool> onChanged;
+}