+++ packages/talker_flutter/lib/src/models/models.dart
@@ -1,5 +1,6 @@
+export 'abstract/abstract.dart';
+export 'additional_talker_setting.dart';
 export 'flutter_talker_data_interface.dart';
 export 'flutter_talker_error.dart';
 export 'flutter_talker_exception.dart';
 export 'flutter_talker_log.dart';
-export 'abstract/abstract.dart';