+++ packages/talker_flutter/lib/src/ui/talker_settings/talker_settings.dart
@@ -1,0 +1,1 @@
+export 'talker_settings_bottom_sheet.dart';