+++ packages/talker_bloc_logger/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.0
+- **FEAT**: Add TalkerBlocLoggerSettings for customize settings
+- **FEAT**: Add Talker addon functionality
+- **FEAT**: Upgrade logs formatting
+
+
+## 0.1.1
+- **FEAT**: Update talker version to 2.4.0
+
 ## 0.1.0
 
 Initial release with base **TalkerBlocObserver** implementation